fix: resolve multi-AI reviewed bugs in console menu, packet capture, and web UI
AGY・Codexによるレビューで指摘された以下を修正:
- コンソールメニューのキー表示と実装の不一致(Port/Size)
- pcap_pkthdr構造体のtv_sec型不一致による受信パケット長の誤読
- WebUIのRestore Defaultsが機能しない問題
- /api/statsのJSON未エスケープとロケール依存の数値書式
- パケットサイズ下限がヘッダーサイズ未満で設定可能だった問題
- 状態変更APIのCSRF対策(Origin/Referer検証 + POST化)

+++ src/NetworkEngine.cs
@@ -43,8 +43,11 @@ namespace MMulticastTool
         public struct pcap_addr { public IntPtr next; public IntPtr addr; public IntPtr netmask; public IntPtr broadaddr; public IntPtr dstaddr; }
         [StructLayout(LayoutKind.Sequential)]
         public struct sockaddr_in { public short sin_family; public ushort sin_port; public uint sin_addr; }
+        // Native pcap_pkthdr uses a C `struct timeval { long tv_sec; long tv_usec; }`.
+        // On Windows, C `long` is always 32-bit (LLP64) regardless of process bitness,
+        // so both fields must map to a 4-byte `int`, not an 8-byte C# `long`.
         [StructLayout(LayoutKind.Sequential)]
-        public struct pcap_pkthdr { public long tv_sec; public int tv_usec; public uint caplen; public uint len; }
+        public struct pcap_pkthdr { public int tv_sec; public int tv_usec; public uint caplen; public uint len; }
         #endregion
 
         public NetworkEngine(AppConfig config, AppStats stats, Action<string, string> logCallback)
